svn merge -r 34105:34228 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@34237 260f80e4-7a28-3924-810f-c04153c831b5

+++ tools/get_archive.py
@@ -288,6 +288,10 @@ def GetFromGsutil(name, directory, version_file, latest_pattern,
       z.close()
     if directory == SDK_DIR:
       unzipped_dir = os.path.join(temp_dir, 'dart-sdk')
+    if os.path.exists(directory):
+      raise Exception(
+          'Removal of directory %s failed. Is the executable running?' %
+          directory)
     shutil.move(unzipped_dir, directory)
   finally:
     shutil.rmtree(temp_dir)
